[squash] Remove required fields from ServerConfig
It's not valid openapi to refer to fields in a subschema pulled in by
`allOf` in the required list. It's also generally not that helpful to
state the fields are required - they happen to always be there with
the current Barman but we don't really need to require them in the
spec. We therefore remove the required list.

Fixes `pg-backup-api serve`.