Allow server names as keys in schema
Updates pg_backup_api.yaml so that the servers object allows any keys.
This fixes the schema by making it align with the `barman diagnose` output
where the keys are the names of the servers.

It is also necessary to remove `archive_command` and `archive_mode` from the
required fields of the Status subschema because a value of `None` is
considered equivalent to not present by openapi-generator and therefore
`None` values (which `barman diagnose` can include in its output) were
causing exceptions.