Support release and snapshot builds
Adds an `is_release` workflow input which can be either 'Y' for a
release build or any other value for a snapshot build.

This is used to determine the version which will be as follows:

* For release builds, the contents of `pg_backup_api/version.txt`.
* For non-release builds, the contents of `pg_backup_api/version.txt`
  with `-snapshot` appended.

The final package name will have the `RELEASE_VERSION` appended for
both release and non-release builds.

The `is_release` input is also used to set the `publish` field in the
Cloudsmith push, such that only non-release builds may republish the
same package version.

To make this work the pkg.sh script is updated to accept the version
from the CLI rather than also obtaining it from
`pg_backup_api/version.txt` directly.

Note that the `snapshot` postfix is dash-separated in the deb package
name but underscore-separated in the rpm package name since dashes
denote separation between the `version` and `release` fields in rpm
names.