Add release process documentation
Adds a README.md to the packaging directory which documents the release
process. The process contains a few steps we could trivially automate
however it is intentionally manual for now so that we can make sure all
the right things are happening - once we're happy we can automate the
things.