Add --directories to fpm invocation for rpms
Adds `--directories /usr/bin/pgbapi-env` to the fpm command for building
rpms. This is needed because simply specifying the virtualenv prefix
alone is not enough for the contents to be removed when the package is
removed.

This is not needed for deb packages where the files are removed
correctly upon package removal.