Ensure unique package names per distro
Renames the generated deb and rpm packages before uploading to
Cloudsmith.

Although we can rely on Cloudsmith tags to differentiate packages this
is only helpful when packages reside in Cloudsmith. Because these
packages are synced elsewhere it is therefore necessary to ensure the
file names are unique for each target distribution.