Add `format: int64` to appropriate integers
Adds the `format: int64` attribute to integers which are likely to
exceed 32 bits (even if the likelihood seems remote at present).

This corrects a bug in the schema where it was describing the size of
backups in bytes as a 32 bit int - this is obviously wrong because it
sets a limit of ~2GB on backup sizes. Other values are less obviously
wrong but could certainly be wrong in some situations - for example a
2GB WAL segment size is unusual but not impossible so we should be using
int64 there.

Values such as counts and times are left as the default int32.

Closes #45.