Remove connexion
Removes the connexion dependency and replaces it with native Flask
routes which just return whatever json `barman.diagnose.exec_diagnose`
produces.

Now that we are no longer packaging this in a venv we can depend on
older versions of Barman since the Barman version on the system will
be the same as the Barman version used as a library by pg-backup-api.
This means it will no longer be necessary to co-ordinate pg-backup-api
and barman upgrades and pg-backup-api will always report the version
of the system barman.

The minimum Barman version is 2.19 though because pg-backup-api depends
on the change to the timestamp formats in the diagnose output which
landed in 2.19.

Closes #48.
Closes #43.
Closes #24.