run.py: Fixing error message when running with no args.
Without this, we've got an usability issue. In the command line you
would see an error like the one below. Because we're using multiple
parsers, we need to check if the `func` attr is defined before
calling it:

Traceback (most recent call last):
  File "/usr/bin/pg-backup-api", line 11, in <module>
    load_entry_point('pg-backup-api==1.1.0', 'console_scripts', 'pg-backup-api')()
  File "/usr/lib/python3.6/site-packages/pg_backup_api/run.py", line 129, in main
    return args.func(args)
AttributeError: 'Namespace' object has no attribute 'func'