Moving LICENSE and README to root dir.
Without this it looks like we're not providing neither of those files,
however they've been there for quite a while, yet, it's easy to not
find them when it's expected to see them right away.

Even github doesn't find it and says this every time you go to check the
code: "Help people interested in this repository understand your project
by adding a README.".