Supporting barman shortcuts.
So far pg-backup-api was calling get_backup() which always receives
the `backup_id` as argument. In order to support barman shortcuts
like "latest" for instance, we need to find out the backup_id first
and call get_backup() afterwards.

This step was not considered beforehand when we released pg-backup-api
v1.1 but customers are already expecting to use such feature with
barman.