Refactor setup_logging function.
We don't need to call this function from the start. One reason is if we
do, we need to check that we run the program as the `barman` user but
if we just need to check a job we sent for instance, this check is not
really required.

We can query the endpoints using plan text requests with curl. So for
now, the only time when we need the `barman` user is when gunicorn is
started because the logs are going to be written to a barman owned file,
hence rw access is needed.

Also, the function was renamed to setup_logging_for_wsgi_server().

This traceback is shown without this change:
 Traceback (most recent call last):
  File "/usr/lib64/python3.6/logging/config.py", line 565, in configure
    handler = self.configure_handler(handlers[name])
  File "/usr/lib64/python3.6/logging/config.py", line 738, in configure_handler
    result = factory(**kwargs)
  File "/usr/lib64/python3.6/logging/__init__.py", line 1032, in __init__
    StreamHandler.__init__(self, self._open())
  File "/usr/lib64/python3.6/logging/__init__.py", line 1061, in _open
    return open(self.baseFilename, self.mode, encoding=self.encoding)
PermissionError: [Errno 13] Permission denied: '/var/log/barman/barman-api.log'