WIP: Add a way of creating and executing "config switch" operations
This commit is a WIP -- the final implementation in the ``pg-backup-api``
depends on the final implementation of the ``barman config switch``
command in Barman.

Through this commit we create a draft for the future change in
``pg-backup-api``. Currently it is composed by the following changes:

* Create a ``ConfigSwitchOperation`` class, which takes care of
validating the arguments for a config switch operation, and also
of running such operation;
* Modify `servers_operations_post` Flask end point so it is able
to receive POST requests both for creating a recovery operation
or a config switch operation;
* Create a new command ``pg-backup-api config-switch``, which is
used by the API in order to run the config swith on Barman;
* Create a function ``config_switch_operation``, which is used by
``pg-backup-api config-switch`` command;
* Create a helper function ``_run_operation``, which is used both
by ``recovery_operation`` and ``config_switch_operation`` functions.
They both have a very similar logic, so that helper function executes
the common code paths.

References: BAR-125.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>