Add missing unit tests
This commit adds an extensive set of unit tests to ensure all
flask endpoints are working as expected. We are mainly interested
in the HTTP status code and in the returned response bodies. In
any case the unit tests are checking also that the underlying
objects and methods are instantiated and called as expected.

Besides that, we also do the following:

* Add unit tests for all utilitary functions
* Add unit tests for all functions used by the CLI
* Add unit tests for the CLI

References: BAR-132.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>