Add `tests.yml` workflow
This commit creates the `tests.yml` workflow. It is responsible
for running all the tests which are exposed by `tox.ini`:

* Lint with `flake8`
* Check dependencies with `pipdeptree`
* Unit tests and coverage with `pytest`
* Static type checking with `pyright`

`flake8` and `pipdeptree` run only once, with Ubuntu and with the
latest Python version.

On the other hand, the unit tests and static type checkings are
run using a matrix of operating systems and Python versions.

References: BAR-133.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>