WIP: Adjust unit tests and add missing unit tests
After #87 has been merged into `main`, this dev branch for BAR-125
has been rebased on top of `main`.

Given the improvements made to the unit test suite through #87,
we are now adjusting the unit tests of this dev branch:

* `test_server_operation.py` for `ConfigSwitchOperation` is now
  following the new design, which uses `pytest` and splits complex
  tests into several smaller ones;
* add tests for "config switch operation" on unit test modules that
  where introduced by #87 (`test_main.py`, `test_run.py`, and
  `test_utility_controller.py`)

Note: we are still missing the final design of the `barman config switch`
implementation in Barman, so we keep this branch as a WIP.

References: BAR-125.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>