Add workflow for performing tests on pushes and PRs (#88)
This PR adds a `tox.ini` file to the repository. It exposes a
few different test environments which can be used by the developer
for:

* Linting the code
* Checking Python modules dependencies
* Running unit tests with coverage report
* Running a static type checker

Instructions on how to use `tox` have been added to the `README.md`
file.

These facilities exposed by the implemented `tox.ini` file are used
by the `tests.yml` workflow which is also introduced by this PR. It is
responsible for running all the tests which are exposed by `tox.ini`

As part of this PR we also applied a few fixes which were reported by
the workflow runs:

* Issues reported by `flake8` in `setup.py`
* Issues reported by `pyright` in `server_operation.py`, `utils.py`,
  and `utility_controller.py`
* Issues on unit test `test_main_helper`: it was failing because of a
  difference between the terminal size of runners and our laptops
* Issues on unit test `test_server_operation_post_not_json` when
  ran through GitHub Actions with Python 3.7 + Flask 2.2.5. That
  combination caused Flask to return `400 Bad Request` instead of
  `415 Unsupported Media Type`, which is returned by other
  combinations

References: BAR-133.