Adjust code and unit tests to reflect the final design of `barman config-switch`
Previous to this commit we had draft code and unit tests as we were still
working on the implementation of `barman config-switch` command.

Now that the design of `barman config-switch` is mature, we are adjusting
this branch so it gets compliant with the expected implementation of that
command.

References: BAR-125.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>