Add support for Barman instance operations
So far we had only operations related with Barman servers in the
API:

* `RecoveryOperation`: to perform a `barman recover` of a Barman
  server
* `ConfigSwitchOperation`: to perform a `barman config-switch` to
  a Barman server

However, we intend to add an operation to the API which is performed
at the Barman instance level (global), not to a specific Barman
server: the `ConfigModelOperation`.

With that in mind this initial commit changes the classes below so
we will be able to create Barman instance operations:

* `OperationServer`: turn `server_name` argument optional. When it
  is `None`, that is considered an instance operation. Thus, the
  `job` and `output` files will be written under the Barman home
  in that case instead of under a Barman server directory
* `Operation`: turn `server_name` argument optional, so it creates
  `OperationServer` accrodingly

Unit tests changed accordingly, so we check both server and instance
operations.

Follow-up commits to come, which will implement new endpoints to the
API and the `ConfigModelOperation`.

References: BAR-126.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>