Add endpoints for handling Barman instance operations
So far we had only endpoints for handling Barman server operations. This
commit is a preparation step before implementing the first instance operation.

As part of the work we added a couple new endpoints to the API:

* `/operations/<operation_id>`: used to get the status of an instance
  operation -- similar to what we have for server operations through
  `/servers/<server_name>/operations/<operation_id>`;
* `/operations`: used to get a list of or instance operations, or to create
  a new instance operation -- similar to what we have for server operations
  through `/servers/<server_name>/operations`.

The handling of `POST` requests to `/operations` is only returning a dummy
message for now, as we still need to implement the first instance operation.

Unit tests were modified accordingly.

References: BAR-126.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>