Rename "config model" as "config update", and adjust the missing logic
We initially designed a command in Barman called `barman config model`
to create and update models.

However, as part of BAR-128 we ended up implementing `barman config-update`
which is able to handle both server and models.

This commit reflects that change in the name of the command, and also
implements the actual logic for listening on the pg-backup-api endpoint
and properly calling the Barman command.

References: BAR-126.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>