Please `pyright`
`pyright` was complaining about the `exec_diagnose`. That issue occurs because
based on the Barman module version that is installed in the virtual
environemnt, the signature of that method would have either 2 or 4 parameters.

We had 2 options to handle that issue:

* Add `pyright: ignore` directives to the source code where `exec_diagnose` is
  called; or
* Change the implementation so instead of relying on `AttributeError`
  exception, it would handle the version for the Barman module.

We chose the first option for now, so we unblock the release. The code with
`AttributeError` exception has already been tested. If we were to go with the
second option now, that would require changing the code, changing the unit tests
and performing more tests, thus dealying the release.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>