Remove flask as dependency
pg-backup-api is a simple, low traffic WSGI application that can be
rewritten for "wsgiref" compatibility instead of "flask" microframework.
At the moment, Flask is not available in RHEL8, and despite all others
distros already include this module in the different repositories,
migrate to a basic *built-in* python3 module sounds sensible enough to
start making this port.

Whith this change, we're relying on a module that it's updated with
every release of python as well. We had some small dependance on "flask"
that made this call simpler after a careful consideration.